% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{adjacency}
\alias{adjacency}
\title{Find adjacent tokens in a corpus}
\usage{
adjacency(corpus_data, skip = 0)
}
\arguments{
\item{corpus_data}{a corpus object from created with quanteda::corpus()}

\item{skip}{integer, the number of tokens to skip to calculate adjacency}
}
\value{
a data frame showing the skip-grams of the input corpus with columns "feature" "frequency" and "gram"
}
\description{
\code{adjacency} extracts adjacent tokens for a user definded skip-gram from
a text corpus and calculates their frequency using the quanteda package. Skip-grams are adjacent words, where a word in between is skipped.
Example: "quality life" is a skip-1-gram of "quality of life".
}
\examples{
\dontrun{
adjacency(corpus)
}

}
