% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Run App  server logic}
\usage{
app_server(input, output, session)
}
\value{
a shiny app
}
\description{
Run App  server logic
}
\examples{
shinyApp(ui = app_ui, server = app_server)

}
