% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\docType{data}
\name{app_ui}
\alias{app_ui}
\title{UI logic for run_app()}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 5.
}
\usage{
app_ui
}
\value{
the UI for a shiny app
}
\description{
UI logic for run_app()
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(shiny)
shinyApp(ui = app_ui, server = app_server)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
