% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_MeSH_norm_frequencies.R
\name{calculate_mesh_norms}
\alias{calculate_mesh_norms}
\title{Calculate the MeSH norms for a preprocessed population set}
\usage{
calculate_mesh_norms(dictionary, population_mesh)
}
\arguments{
\item{dictionary}{a MeSH dictionary created with \code{update_mesh}}

\item{population_mesh}{a population set preprocessed with \code{prepare_MeSH_table}}
}
\value{
A data frame with the norm frequencies.
}
\description{
\code{calculate_mesh_norms} calculates the probabilities of MeSH terms.
}
\examples{
\dontrun{
calculate_mesh_norms (MeSH_dictionary, populationSet)}
}
