% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_corpus.R
\name{create_corpus}
\alias{create_corpus}
\title{Create a quanteda text corpus from bibliographic data}
\usage{
create_corpus(reference_set)
}
\arguments{
\item{reference_set}{a list object created with read_bibliography()}
}
\value{
a quanteda text corpus with titles and abstracts as the text corpus
}
\description{
\code{create_corpus} accepts an list object of class bibliography and
extracts the titles and abstracts for a text analysis.
Empty titles and abstracts are marked with "NO_TITLE" or "NO_ABSTRACT" in the corpus.
Author names are normalized and problems with missing first authors are handled by substituting empty first author fields with "Anonymous".
}
\examples{
\dontrun{
create_corpus(references)
}
}
