% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_MeSH_norm_frequencies.R
\name{create_MeSH_norm_frequencies}
\alias{create_MeSH_norm_frequencies}
\title{Calculate new probabilities for MeSH and qualifier for a given population set}
\usage{
create_MeSH_norm_frequencies(reference_set, mesh_xml = NULL, qual_xml = NULL)
}
\arguments{
\item{reference_set}{a reference_set in RIS or PubMed format created with read_bibliography()}

\item{mesh_xml}{optional argument, character vector with the path to an xml file containing MeSH descriptors in the NLM format}

\item{qual_xml}{optional argument, character vector with the path to an xml file containing MeSH descriptors in the NLM format}
}
\value{
a list of 2 data frames with MeSH ("headings") and "qualifier" norms and two list with raw MeSH headings and MeSH qualifiers
}
\description{
Calculate new probabilities for MeSH and qualifier for a given population set
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ris <- c("TY  - JOUR",
        "AU  - Kapp",
        "TI  - Titles",
        "PY  - 2023",
        "JOUR  - IQWiG Journal",
        "KW  - Systematic Reviews as Topic",
        "ER  -")

tmp <- tempfile(fileext = ".txt")
writeLines(ris, tmp)
reference_set <- read_bibliography(tmp, return_df = FALSE)
create_MeSH_norm_frequencies(reference_set)
\dontshow{\}) # examplesIf}
}
