% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_scores.R
\name{create_MeSH_qualifier_lists}
\alias{create_MeSH_qualifier_lists}
\title{Utilities function to merge testset and population set data.}
\usage{
create_MeSH_qualifier_lists(testset)
}
\arguments{
\item{testset}{an object created with create_testset() to be used fo analysis instead of a raw ris file}
}
\value{
the input object with processed data
}
\description{
The function extracts the MeSH terms and qualifiers from the testset data and counts their frequency.
}
\examples{
\dontrun{
create_MeSH_qualifier_lists(testset)
}
}
