% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_popset.R
\name{create_popset}
\alias{create_popset}
\title{Process raw Endnote ris file with PubMed data into population norm set}
\usage{
create_popset(risfile, update_MeSH = FALSE, mesh_xml = NULL, qual_xml = NULL)
}
\arguments{
\item{risfile}{a ris-format file to be imported}
}
\value{
popset a list of objects
popset
...$freetext
...$MeSH.terms
...$qualifier
}
\description{
Process raw Endnote ris file with PubMed data into population norm set
}
\examples{
ris <- c("TY  - JOUR",
         "AU  - Kapp",
         "TI  - Titles",
         "PY  - 2023",
         "JOUR  - IQWiG Journal",
         "KW  - Systematic Reviews as Topic",
         "ER  -")

tmp <- tempfile(fileext = ".txt")
writeLines(ris, tmp)
create_popset(tmp)
}
