% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_testset.R
\name{create_testset}
\alias{create_testset}
\title{Title}
\usage{
create_testset(risfile, dev_set = FALSE, project_name = TRUE, seed = NULL)
}
\arguments{
\item{risfile}{a ri-format file to be imported}

\item{dev_set}{should a random development set (2/3 imported references) be created}

\item{project_name}{should the project name be guessed from the file name}

\item{seed}{option to set a random seed, if dev_set = TRUE}
}
\value{
a list in the format

testset
...$freetext
...$MeSH.terms / ...$all_keywords / ...$MeSH_with_Qualifier
...$PMIDS / ...$testset | ...$devevolpment_set /...$validation_set
...$reference.list
...$text_corpus
if (dev_set = TRUE) ...$seed

#' @examples
ris <- c("TY  - JOUR",
"AU  - Kapp",
"TI  - Titles",
"PY  - 2023",
"JOUR  - IQWiG Journal",
"KW  - Systematic Reviews as Topic",
"ER  -")

tmp <- tempfile(fileext = ".txt")
writeLines(ris, tmp)
create_testset(tmp)
}
\description{
Title
}
