% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_popset.R
\name{create_UID_esearch_terms}
\alias{create_UID_esearch_terms}
\title{Create a valid PubMed search query using random integer numbers as PMIDs}
\usage{
create_UID_esearch_terms(sample_size = 20000, max_pmid = 4e+07, seed = NULL)
}
\arguments{
\item{sample_size}{positive integer value < 100,000 defining the size of references to be retrieved}

\item{max_pmid}{the highest PMID to be considered, defaults to 40000000.}

\item{seed}{an integer value to control the random seed used for sampling}
}
\value{
A list of character vectors each containing a valid PubMed search query of a maximum length of 20000 character symbols.
Each list value can be passed on to the "term" argument of esearch_post_req() to create a POST request.
}
\description{
Create a valid PubMed search query using random integer numbers as PMIDs
}
\examples{
create_UID_esearch_terms(sample_size = 4, max_pmid = 30000000)

}
