% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_popset.R
\name{efetch_with_idlist}
\alias{efetch_with_idlist}
\title{Creating an Efetch POST request for PubMed via EUtils API from a list of UIDs (PMIDs)}
\usage{
efetch_with_idlist(idlist, email)
}
\arguments{
\item{idlist}{list of PMIDs}

\item{email}{a valid e-mail address, which will be sent with the API request to identify the API user (required by NLM)}
}
\value{
a character vector containing the requested PubMed reference metadata in MEDLINE format
}
\description{
Creating an Efetch POST request for PubMed via EUtils API from a list of UIDs (PMIDs)
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
validEmail <- "ENTER_VALID_EMAIL"
efetch_with_idlist(idlist = list( 22587829), email = validEmail)
\dontshow{\}) # examplesIf}
}
