% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_popset.R
\name{efetch_with_queryKey}
\alias{efetch_with_queryKey}
\title{Creating an Efetch GET request via E-Utilities using the history server}
\usage{
efetch_with_queryKey(queryKey, webEnv, email)
}
\arguments{
\item{queryKey}{Integer or character value containing an integer query key returned by a previous ESearch}

\item{webEnv}{a character vector providing a Web environment string returned from a previous ESearch (see also https://www.ncbi.nlm.nih.gov/books/NBK25499/#chapter4.ESearch)}

\item{email}{a valid e-mail address, which will be sent with the API request to identify the API user (required by NLM)}
}
\value{
a character vector containing the requested PubMed reference metadata in MEDLINE format
}
\description{
Creating an Efetch GET request via E-Utilities using the history server
}
\examples{
\dontrun{
efetch_with_queryKey(queryKey = 1, webEnv = "JFm5NJfsp6rHJGKsdjg576GU" , email = "xx@xx.xxxxx")
}

}
