% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_popset.R
\name{esearch_post_req}
\alias{esearch_post_req}
\title{Creating an esearch POST request via E-Utilities using the history server}
\usage{
esearch_post_req(
  term,
  tool = "searchbuildR",
  email,
  retmax = NULL,
  webEnv = NULL
)
}
\arguments{
\item{term}{a character vector with a valid Pubmed query (see also https://pubmed.ncbi.nlm.nih.gov/help/#how-do-i-search-pubmed)}

\item{tool}{a character vector providing information about the tool that sends the request (defaults to package name "searchbuildR")}

\item{email}{a valid e-mail address, which will be sent with the API request to identify the API user (required by NLM)}

\item{retmax}{the maximum number of references to be retrieved from the total search result}

\item{webEnv}{a character vector providing a Web environment string returned from a previous ESearch (see also https://www.ncbi.nlm.nih.gov/books/NBK25499/#chapter4.ESearch)}
}
\value{
an HTTP POST request which can be sent to the E-Utilities API
}
\description{
Creating an esearch POST request via E-Utilities using the history server
}
\examples{
searchbuildR:::esearch_post_req(term="20204[pdat]",email="xx@xx.xxxx")
#|>  httr2::req_perform()

}
