% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_mesh_entry_terms.R
\name{get_mesh_entry_terms}
\alias{get_mesh_entry_terms}
\title{Retrieve MeSH entry terms for MeSH Headings}
\usage{
get_mesh_entry_terms(meshTerm)
}
\arguments{
\item{meshTerm}{a character vector with MeSH headings}
}
\value{
a list of entry terms
}
\description{
Retrieve MeSH entry terms for MeSH Headings
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_mesh_entry_terms(c("Breast Neoplasms", "Gingivitis"))
\dontshow{\}) # examplesIf}
}
