% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MeSH_Dictionary}
\alias{MeSH_Dictionary}
\title{PubMed MeSH Headings 2022}
\format{
\subsection{\code{MeSH_Dictionary}}{

A data frame with 3094 rows and 1 column:
\describe{
\item{MeSH}{MeSH Heading}
}
}
}
\source{
\url{https://www.nlm.nih.gov/databases/download/mesh.html}
}
\usage{
MeSH_Dictionary
}
\description{
A comprehensive list of the MeSH headings as data.frame
}
\keyword{datasets}
