% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_esearch.R
\name{mesh_esearch}
\alias{mesh_esearch}
\title{Search MeSH Terms in MeSH tree}
\usage{
mesh_esearch(meshTerm)
}
\arguments{
\item{meshTerm}{a character vector with MeSH headings}
}
\value{
a character vector with MeSH unique identifiers
}
\description{
Search MeSH Terms in MeSH tree
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mesh_esearch(c("Breast Neoplasms", "Gingivitis"))
\dontshow{\}) # examplesIf}
}
