% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_esummary.R
\name{mesh_esummary}
\alias{mesh_esummary}
\title{Get MeSH entry terms from MeSH tree}
\usage{
mesh_esummary(uids)
}
\arguments{
\item{uids}{a vector with unique MeSH identifiers}
}
\value{
a list of entry terms and MeSH parent subheadings
}
\description{
Get MeSH entry terms from MeSH tree
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mesh_esummary(c("68005891","68001943"))
\dontshow{\}) # examplesIf}
}
