% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_popset.R
\name{plot_popset_data}
\alias{plot_popset_data}
\title{Plot the distribution of publication years for a search result}
\usage{
plot_popset_data(result, seed)
}
\arguments{
\item{result}{character vector, MEDLINE format search result}

\item{seed}{a random seed used to sample random PMIDs}
}
\value{
a histogram
}
\description{
Plot the distribution of publication years for a search result
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
result <- c("PMID- 1",
"DP  - 2021",
"PMID- 2",
"DP  - 2023",
"PMID- 3",
"DP  - 2022")
plot_popset_data(result = result, seed = 1)
\dontshow{\}) # examplesIf}
}
