% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popsetRCT}
\alias{popsetRCT}
\title{PubMed random population set with PuBMed RCT filter}
\format{
\subsection{\code{popsetRCT}}{

A character vector of length 10092
}
}
\source{
\url{https://www.pubmed.com}
}
\usage{
popsetRCT
}
\description{
A random sample of 100 PubMed references that match a search : 2000:2024[pdat] AND therapy/broad[filter],
search date: July 31st 2024. Source: National Library of Medicine.
}
\keyword{datasets}
