% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_popset.R
\name{post_random_PMIDs}
\alias{post_random_PMIDs}
\title{Send a POST request with mulitple a search lines to PubMed via Eutils API.}
\usage{
post_random_PMIDs(UIDSearchTermList, sample_size, webEnv, email)
}
\arguments{
\item{UIDSearchTermList}{a list of character strings which are valid PubMed search queries.}

\item{sample_size}{the number of references to be retrieved}

\item{webEnv}{a character vector providing a Web environment string returned from a previous ESearch (see also https://www.ncbi.nlm.nih.gov/books/NBK25499/#chapter4.ESearch)}

\item{email}{a valid e-mail address, which will be sent with the API request to identify the API user (required by NLM)}
}
\value{
a list containing countHits(the last search result), "queryKey"(all search line indices) and "counter" (total search size result size)
}
\description{
\code{post_random_PMIDs} takes a list of search queries and combines each with (#1 AND), POSTs it to the PubMed history server
and retrieves the search result count. The first search (#1) in the current search history (webEnv) thus
will be used as a filter.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
searchTermList <- list("1", "2")
validwebEnv <-"VALID_WEB_ENV"
validEmail <- "ENTER_VALID_EMAIL"
post_random_PMIDs(searchTermList, sample_size = 2, webEnv = validwebEnv, email = validEmail)
\dontshow{\}) # examplesIf}
}
