% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_freq_table.R
\name{prepare_freq_table}
\alias{prepare_freq_table}
\title{Preprocess the titles and abstracts for text analysis}
\usage{
prepare_freq_table(corpus_ref)
}
\arguments{
\item{corpus_ref}{a corpus object}
}
\value{
a processed corpus object
}
\description{
remove all special characters
see also:    https://de.wikipedia.org/wiki/Liste_der_Unicode-Eigenschaften
punctuation: remove all characters in the Unicode "Punctuation" P class⁠
symbols:     all characters in the Unicode "Symbol" S⁠ class
hyphens:     if FALSE, do not split words that are connected by hyphenation and hyphenation-like characters in between words, e.g. "self-aware" becomes c("self", "-", "aware")
numbers:     remove tokens that consist only of numbers, but not words that start with digits, e.g. ⁠2day⁠
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
corpusObj <- quanteda::corpus("some text")
prepare_freq_table(corpusObj)
\dontshow{\}) # examplesIf}
}
