% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_MeSH_table.R
\name{prepare_MeSH_table}
\alias{prepare_MeSH_table}
\title{Extract and count the keywords from a RIS file}
\usage{
prepare_MeSH_table(reference_set)
}
\arguments{
\item{reference_set}{a list object created with read_bibliography()}
}
\value{
a list of 3 lists of unique keywords containing \verb{all keywords}, MeSH/qualifier combinations (\code{MeSH_with_qualifier}) or major headings (\code{MeSH_starred})
}
\description{
Extract and count the keywords from a RIS file
}
\examples{
\dontrun{
references <- create_testset(ris)
prepare_MeSH_table(references)}
}
