% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Qualifier_Dictionary}
\alias{Qualifier_Dictionary}
\title{PubMed Qualifier 2022}
\format{
\subsection{\code{Qualifier_Dictionary}}{

A data frame with 76 rows and 1 column:
\describe{
\item{qualifier}{MeSH qualifier}
}
}
}
\source{
\url{https://www.nlm.nih.gov/databases/download/mesh.html}
}
\usage{
Qualifier_Dictionary
}
\description{
A comprehensive list of the MeSH qualifier as data.frame
}
\keyword{datasets}
