% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_fetch.R
\name{random_pubmed_fetch}
\alias{random_pubmed_fetch}
\title{Fetch random sample of pubmed references}
\usage{
random_pubmed_fetch(size, maxUID)
}
\arguments{
\item{size}{approximate number of references to be sampled}

\item{maxUID}{biggest UID to be considered for random sample}
}
\value{
a list with two elements `seed´ (the random seed applied) and ´result´ (retrieved references in PubMed format)
}
\description{
Fetch random sample of pubmed references
}
\examples{
searchbuildR:::random_pubmed_fetch(size = 2, maxUID = 37000000)
}
