% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{raw_ris_data}
\alias{raw_ris_data}
\title{Parse a RIS or PubMed format text file into a list}
\usage{
raw_ris_data(x)
}
\arguments{
\item{x}{path to a RIS or PubMed format file}
}
\value{
a list with each reference as a character vector
}
\description{
Parse a RIS or PubMed format text file into a list
}
\examples{
ris <- c("TY  - JOUR",
        "AU  - Kapp",
        "TI  - Titles",
        "PY  - 2023",
        "JOUR  - IQWiG Journal",
        "KW  - Systematic Reviews as Topic",
        "ER  -")

tmp <- tempfile(fileext = ".txt")
writeLines(ris, tmp)
raw_ris_data(tmp)

}
