% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bibliography.R
\name{read_bibliography}
\alias{read_bibliography}
\title{import a ris file as class bibliography}
\usage{
read_bibliography(filename, return_df = TRUE)
}
\arguments{
\item{filename}{a risfile to import}

\item{return_df}{should a dataframe be returned, if FALSE returns a list}
}
\value{
an object of class bibliography from the revtools package
}
\description{
import a ris file as class bibliography
}
\examples{
ris <- c("TY  - JOUR",
        "AU  - Kapp",
        "TI  - Titles",
        "PY  - 2023",
        "JOUR  - IQWiG Journal",
        "KW  - Systematic Reviews as Topic",
        "ER  -")

tmp <- tempfile(fileext = ".txt")
writeLines(ris, tmp)
read_bibliography(tmp)
}
