% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_pmids}
\alias{return_pmids}
\title{Extract the Accession Numbers from a bibliography class object}
\usage{
return_pmids(testset_ref, random_references = NULL, validation_set = FALSE)
}
\arguments{
\item{testset_ref}{a list of class bibliography}

\item{random_references}{a vector of indices of testset_ref which will be the development set}

\item{validation_set}{should the accession numbers of the validation set be returned as well}
}
\value{
A list with all accession numbers \code{testset} and optionally the accession numbers for \code{development_set} and \code{validation_set}
}
\description{
Extract the Accession Numbers from a bibliography class object
}
\examples{
\dontrun{
references <- read_bibliography(ris, return_df = FALSE)
return_pmids(references)
}
}
