% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Runs the shiny app of the package}
\usage{
run_app()
}
\value{
a shiny app
}
\description{
Runs the shiny app of the package
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

run_app()
#see vignette
\dontshow{\}) # examplesIf}
}
