% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchbuilder-package.R
\docType{package}
\name{searchbuilder-package}
\alias{searchbuilder}
\alias{searchbuilder-package}
\title{searchbuilder: Identify Search Terms in Evidence Synthesis}
\description{
Currently in development: R Package to identify search terms in evidence synthesis with quantitative textual analysis.
}
\author{
\strong{Maintainer}: Claudia Kapp \email{claudia.kapp@iqwig.de} (\href{https://orcid.org/0000-0002-7111-1897}{ORCID})

}
\keyword{internal}
