% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchterms-package.R
\docType{package}
\name{searchterms-package}
\alias{searchterms}
\alias{searchterms-package}
\title{searchterms: Identify Search Terms in Evidence Synthesis}
\description{
Currently in development: R Package to identify search terms in evidence synthesis with quantitative textual analysis.
}
\author{
\strong{Maintainer}: Claudia Kapp \email{claudia.kapp@iqwig.de} (\href{https://orcid.org/0000-0002-7111-1897}{ORCID})

}
\keyword{internal}
