% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{summarise_adjacency}
\alias{summarise_adjacency}
\title{A wrapper function to summarise adjacency() for multiple skip-grams}
\usage{
summarise_adjacency(corpus_data, ngrams = 4)
}
\arguments{
\item{corpus_data}{a corpus object from created with quanteda::corpus()}

\item{ngrams}{integer, defines the window of tokens (ngram) in which to identify skip-grams}
}
\value{
a data frame showing the skip-grams of the input corpus with columns "feature" "frequency"
and a column "Skip-x" for the frequency of each skip-gram.
}
\description{
A wrapper function to summarise adjacency() for multiple skip-grams
}
\examples{
\dontrun{
summarise_adjacency(corpus)
}
}
