% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testset}
\alias{testset}
\title{Testset example}
\format{
\subsection{\code{testset}}{

a list of 4 data frames
\describe{
freetext  S3: ‘frequency’, ‘textstat’ and 'data.frame':	172 rows of  16 columns:
\describe{
\item{feature}{character, single word token/term}
\item{frequency}{double, observed frequency of the single word token/term}
\item{group}{"all"}
\item{n}{sum of all terms included in analysis}
\item{project}{name of the import file}
\item{rank}{integer, rank of the term according to observed frequency}
\item{Norm.frequency}{numeric, absolute frequency of each token/term in population set}
\item{Norm.rank}{numeric, rank of token/term in population set}
\item{Norm.docfreq}{numeric, absolute document frequency of each token/term in population set}
\item{N}{numeric, total frequency of all token/terms in population set}
\item{p}{estimated proportional frequency based on Norm.frequency}
\item{E}{expected value}
\item{var}{estimated variance}
\item{z}{z-score of bionomial test}
\item{approx_criteria}{}
}
MeSH.Terms:S3:tbl_df, tbl, data.frame :	8 rows and 13 columns:
\describe{
\item{MeSH}{character, MeSH heading title style case}
\item{frequency}{integer, observed frequency of MeSH term}
\item{docfreq}{integer, observed frequency of the MeSH term}
\item{project}{name of the import file}
\item{n}{sum of all MeSH-terms included in analysis}
\item{Norm.docfreq}{numeric, absolute document frequency of each MeSH heading in population set}
\item{Norm.frequency}{numeric, absolute frequency of each MeSH heading in population set}
\item{N}{numeric, total frequency of all MeSH headings in population set}
\item{p}{numeric, p-value of binomial test of Norm.frequency}
\item{E}{expected value}
\item{var}{estimated variance}
\item{z}{z-score of bionomial test}
\item{approx_criteria}{}
}
qualifier :S3:tbl_df, tbl, data.frame :	4 rows and 13 columns:
\describe{
\item{qualifier}{character, MeSH qualifier all lower case}
\item{frequency}{integer, observed frequency of MeSH term}
\item{docfreq}{integer, observed frequency of the MeSH term}
\item{project}{name of the import file}
\item{n}{sum of all MeSH-terms included in analysis}
\item{Norm.docfreq}{numeric, absolute document frequency of each MeSH heading in population set}
\item{Norm.frequency}{numeric, absolute frequency of each MeSH heading in population set}
\item{N}{numeric, total frequency of all MeSH headings in population set}
\item{p}{numeric, p-value of binomial test of Norm.frequency}
\item{E}{expected value}
\item{var}{estimated variance}
\item{z}{z-score of bionomial test}
\item{approx_criteria}{}
}
\describe{
all_keywords:S3:tbl_df, tbl, data.frame :	12 rows and 4 columns:
\item{MeSH}{character, MeSH heading title style case}
\item{frequency}{integer, observed frequency of MeSH term}
\item{docfreq}{integer, observed frequency of the MeSH term}
\item{project}{name of the import file}
}
\describe{
leftover_keywords: an empty list
}
}
}
}
\usage{
testset
}
\description{
example result of function "z_scores"
}
\keyword{datasets}
