% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_MeSH_norm_frequencies.R
\name{update_mesh}
\alias{update_mesh}
\title{Parsing a MeSH xml}
\usage{
update_mesh(mesh_xml)
}
\arguments{
\item{mesh_xml}{optional argument, character vector with the path to an xml file containing MeSH descriptors in the NLM format}
}
\value{
a list object containing two data frames with the MeSH specific Names, UIs and Tree Numbers
}
\description{
\code{update_mesh} parses an NLM provided XML MeSH file and extracts Names, UIs and Tree Numbers into data frames.
}
\examples{
\dontrun{
update_mesh("path/to/mesh.xml")}
}
