% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Run App  server logic}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{input from server}

\item{output}{output to UI}

\item{session}{session ID}
}
\value{
a shiny app
}
\description{
Run App  server logic
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(shiny)
shinyApp(ui = app_ui, server = app_server)
\dontshow{\}) # examplesIf}
}
