% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_mesh_entry_terms.R
\name{mesh_esearch}
\alias{mesh_esearch}
\title{Search MeSH Terms in MeSH tree}
\usage{
mesh_esearch(meshTerm, email)
}
\arguments{
\item{meshTerm}{a character vector with MeSH headings}

\item{email}{a valid e-mail address, which will be sent with the API request to identify the API user (required by NLM)}
}
\value{
a character vector with MeSH unique identifiers
}
\description{
Search MeSH Terms in MeSH tree
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mesh_esearch(c("Breast Neoplasms", "Gingivitis"), email = "xx@xx.xxxxx")
\dontshow{\}) # examplesIf}
}
