% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchbuildR-package.R
\docType{package}
\name{searchbuildR-package}
\alias{searchbuildR}
\alias{searchbuildR-package}
\title{searchbuildR: Identify Search Terms in Evidence Synthesis}
\description{
R Package to identify search terms in evidence synthesis with quantitative textual analysis.
}
\author{
\strong{Maintainer}: Institut für Qualität und Wirtschaftlichkeit im Gesundheitswesen(IQWiG) \email{recherche@iqwig.de} [copyright holder]

}
\keyword{internal}
