% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchbuildR-package.R
\docType{package}
\name{searchbuildR-package}
\alias{searchbuildR}
\alias{searchbuildR-package}
\title{searchbuildR: Identify Search Terms in Evidence Synthesis}
\description{
R Package to identify search terms in evidence synthesis with quantitative textual analysis.
}
\author{
\strong{Maintainer}: Claudia Kapp \email{claudia.kapp@iqwig.de} (\href{https://orcid.org/0000-0002-7111-1897}{ORCID})

}
\keyword{internal}
