% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_scores.R
\name{z_scores}
\alias{z_scores}
\title{Z score tables}
\usage{
z_scores(
  risfile = NULL,
  references = NULL,
  risfile_population,
  load_popset = TRUE,
  dev_set = FALSE,
  seed = NULL
)
}
\arguments{
\item{risfile}{a ris file containing a testset}

\item{references}{an object created with create_testset() to be used fo analysis instead of a raw ris file}

\item{risfile_population}{a dataframe containing a population set}

\item{load_popset}{logical, should internal population set be applied (default) or should a customized population set be calculated}

\item{dev_set}{logical, if testset should be randomly divided into a development and validation set select TRUE?}

\item{seed}{should be included, to reproduce a created development set from a previous session}
}
\value{
a list
}
\description{
Z score tables
}
\examples{
ris <- c("TY  - JOUR",
         "AU  - Kapp",
         "TI  - Titles",
         "PY  - 2023",
         "JOUR  - IQWiG Journal",
         "KW  - Systematic Reviews as Topic",
         "ER  -")

tmp <- tempfile(fileext = ".txt")
writeLines(ris, tmp)
z_scores(tmp)

}
